chore(chrome-extension): Align with SDK spec
