+++ packages/chrome-extension/package.json
@@ -26,7 +26,10 @@
   "module": "./dist/esm/index.js",
   "types": "./dist/types/index.d.ts",
   "files": [
-    "dist"
+    "background",
+    "dist",
+    "internal",
+    "react"
   ],
   "scripts": {
     "build": "tsup",
