+++ packages/chrome-extension/src/background/index.ts
@@ -1,0 +1,2 @@
+export type { StorageCache } from '../internal/utils/storage';
+export { __unstable__createClerkClient } from './clerk';
