+++ packages/chrome-extension/src/internal/constants.ts
