+++ packages/chrome-extension/src/internal/index.ts
@@ -1,0 +1,1 @@
+export * from './clerk';
