+++ packages/chrome-extension/src/internal/utils/__tests__/cookies.test.ts
