+++ packages/chrome-extension/src/internal/utils/cookies.ts
