+++ packages/chrome-extension/src/internal/utils/errors.ts
