+++ packages/chrome-extension/src/internal/utils/jwt-handler.ts
