+++ packages/chrome-extension/src/internal/utils/manifest.ts
@@ -1,9 +1,9 @@
 import type { SetRequired } from 'type-fest';
 import type { Manifest } from 'webextension-polyfill';
 
+import type { ClerkClientExtensionFeatures } from '../../types';
 import { VALID_HOST_PERMISSION_REGEX } from '../constants';
-import type { ClerkClientExtensionFeatures } from '../types';
-import { errorThrower, missingManifestKeyError, missingValidManifestHostPermission } from '../utils/errors';
+import { errorThrower, missingManifestKeyError, missingValidManifestHostPermission } from './errors';
 
 export type ValidatedManifest = SetRequired<Manifest.WebExtensionManifest, 'permissions' | 'host_permissions'>;
 
@@ -11,22 +11,26 @@ export function validateManifest(
   manifest: Manifest.WebExtensionManifest,
   features: ClerkClientExtensionFeatures,
 ): asserts manifest is ValidatedManifest {
-  const hasFeatures = Object.keys(features).length > 0;
-
-  if (!hasFeatures) {
-    return;
-  }
+  const hasFeatures = Boolean(features) && Object.keys(features).length > 0;
 
   if (!manifest.permissions) {
     return errorThrower.throw(missingManifestKeyError('permissions'));
   }
 
-  if (features.sync && !manifest.permissions.includes('cookies')) {
-    return errorThrower.throw(missingManifestKeyError('permissions.cookies'));
+  if (!manifest.permissions.includes('storage')) {
+    return errorThrower.throw(missingManifestKeyError('permissions.storage'));
   }
 
-  if ((features.background || features.sync) && !manifest.permissions.includes('storage')) {
-    return errorThrower.throw(missingManifestKeyError('permissions.storage'));
+  if (!hasFeatures) {
+    return;
+  }
+
+  if (features.background && !manifest.background) {
+    return errorThrower.throw(missingManifestKeyError('background'));
+  }
+
+  if (features.sync && !manifest.permissions.includes('cookies')) {
+    return errorThrower.throw(missingManifestKeyError('permissions.cookies'));
   }
 
   if (features.sync && !manifest.host_permissions) {
