+++ packages/chrome-extension/src/internal/utils/storage.ts
