+++ packages/chrome-extension/src/react/ClerkProvider.tsx
@@ -3,30 +3,23 @@ import type { ClerkProviderProps as ClerkReactProviderProps } from '@clerk/clerk
 import { ClerkProvider as ClerkReactProvider } from '@clerk/clerk-react';
 import React from 'react';
 
-import { createClerkClient } from './clerk';
-import type { ClerkClientExtensionFeatures } from './types';
-import type { StorageCache } from './utils/storage';
+import { createClerkClient } from '../internal/clerk';
+import type { StorageCache } from '../internal/utils/storage';
 
 type ChromeExtensionClerkProviderProps = ClerkReactProviderProps & {
-  extensionFeatures?: ClerkClientExtensionFeatures;
   storageCache?: StorageCache;
-  /** @deprecated Please use extensionFeatures.sync instead */
   syncSessionWithTab?: boolean;
 };
 
 export function ClerkProvider(props: ChromeExtensionClerkProviderProps): JSX.Element | null {
-  const { children, extensionFeatures = {}, storageCache, syncSessionWithTab, ...rest } = props;
+  const { children, storageCache, syncSessionWithTab, ...rest } = props;
   const { publishableKey = '' } = props;
 
   const [clerkInstance, setClerkInstance] = React.useState<Clerk | null>(null);
 
   React.useEffect(() => {
     void (async () => {
-      if (syncSessionWithTab) {
-        extensionFeatures.sync = true;
-      }
-
-      setClerkInstance(await createClerkClient({ publishableKey, storageCache, extensionFeatures }));
+      setClerkInstance(await createClerkClient({ publishableKey, storageCache, syncSessionWithTab }));
     })();
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
