+++ packages/chrome-extension/src/react/index.ts
@@ -1,0 +1,1 @@
+export { ClerkProvider } from './ClerkProvider';
