+++ packages/chrome-extension/src/types.ts
@@ -1,3 +1,9 @@
+export const SCOPE = {
+  background: 'background',
+} as const;
+
+export type Scope = (typeof SCOPE)[keyof typeof SCOPE];
+
 export interface ClerkClientExtensionFeatures {
   background?: boolean;
   sync?: boolean;
